테스트 코드, 서버 코드 리팩토링, 테스트 DB h2로 변경 및 테스트용 데이터 쿼리 생성

+++ yagieottae_back_end/src/main/java/com/yagieottae_back_end/Entity/Alram.java
@@ -36,7 +36,7 @@ public class Alram
     private Boolean beforeMeal; // true:식전 복용 false: 식후 복용
 
     @NotNull
-    private Long dosingTime; //복용시간
+    private int dosingTime; //복용시간
 
     @NotNull
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
